@@ -1,4 +1,3 @@
-
 {
   "name": "nxt-trendz-cart-features",
   "private": true,
@@ -8,21 +7,24 @@
     "npm": ">=6"
   },
   "dependencies": {
+    "@reduxjs/toolkit": "2.11.2",
     "@testing-library/jest-dom": "5.11.9",
     "@testing-library/react": "11.2.5",
     "@testing-library/user-event": "12.6.2",
+    "axios": "1.13.5",
     "chalk": "4.1.0",
+    "concurrently": "^8.2.2",
     "history": "5.0.0",
     "js-cookie": "3.0.0",
     "msw": "0.33.0",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
+    "nodemon": "^3.1.0",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
     "react-icons": "4.2.0",
     "react-loader-spinner": "4.0.0",
+    "react-redux": "9.2.0",
     "react-router-dom": "5.2.0",
-    "reactjs-popup": "2.0.6",
-    "concurrently": "^8.2.2",
-    "nodemon": "^3.1.0"
+    "reactjs-popup": "2.0.6"
   },
   "proxy": "http://localhost:5000",
   "devDependencies": {
