@@ -2,11 +2,12 @@ import {useState, useEffect} from 'react'
 import {useDispatch, useSelector} from 'react-redux'
 import {Link, useHistory, Redirect} from 'react-router-dom'
 import Cookies from 'js-cookie'
-import {loginUser, resetAuth} from '../../redux/slices/authSlice'
+import {registerUser, resetAuth} from '../../redux/slices/authSlice'
 
 import './index.css'
 
-const LoginForm = () => {
+const RegisterForm = () => {
+  const [name, setName] = useState('')
   const [email, setEmail] = useState('')
   const [password, setPassword] = useState('')
 
@@ -17,13 +18,13 @@ const LoginForm = () => {
   useEffect(() => {
     if (success) {
       dispatch(resetAuth())
-      history.replace('/')
+      history.push('/login')
     }
   }, [success, history, dispatch])
 
   const submitForm = event => {
     event.preventDefault()
-    dispatch(loginUser({email, password}))
+    dispatch(registerUser({name, email, password}))
   }
 
   const jwtToken = Cookies.get('jwt_token')
@@ -32,23 +33,36 @@ const LoginForm = () => {
   }
 
   return (
-    <div className="login-form-container">
+    <div className="register-form-container">
       <img
         src="https://assets.ccbp.in/frontend/react-js/nxt-trendz-logo-img.png"
-        className="login-website-logo-mobile-img"
+        className="register-website-logo-mobile-img"
         alt="website logo"
       />
       <img
         src="https://assets.ccbp.in/frontend/react-js/nxt-trendz-login-img.png"
-        className="login-img"
-        alt="website login"
+        className="register-img"
+        alt="website register"
       />
       <form className="form-container" onSubmit={submitForm}>
         <img
           src="https://assets.ccbp.in/frontend/react-js/nxt-trendz-logo-img.png"
-          className="login-website-logo-desktop-img"
+          className="register-website-logo-desktop-img"
           alt="website logo"
         />
+        <div className="input-container">
+          <label className="input-label" htmlFor="name">
+            NAME
+          </label>
+          <input
+            type="text"
+            id="name"
+            className="input-field"
+            value={name}
+            onChange={e => setName(e.target.value)}
+            placeholder="Name"
+          />
+        </div>
         <div className="input-container">
           <label className="input-label" htmlFor="email">
             EMAIL
@@ -56,7 +70,7 @@ const LoginForm = () => {
           <input
             type="text"
             id="email"
-            className="username-input-field"
+            className="input-field"
             value={email}
             onChange={e => setEmail(e.target.value)}
             placeholder="Email"
@@ -69,22 +83,22 @@ const LoginForm = () => {
           <input
             type="password"
             id="password"
-            className="password-input-field"
+            className="input-field"
             value={password}
             onChange={e => setPassword(e.target.value)}
             placeholder="Password"
           />
         </div>
-        <button type="submit" className="login-button" disabled={loading}>
-          {loading ? 'Logging in...' : 'Login'}
+        <button type="submit" className="register-button" disabled={loading}>
+          {loading ? 'Registering...' : 'Register'}
         </button>
         {error && <p className="error-message">*{error}</p>}
-        <p className="register-link">
-          Don't have an account? <Link to="/register">Register</Link>
+        <p className="login-link">
+          Already have an account? <Link to="/login">Login</Link>
         </p>
       </form>
     </div>
   )
 }
 
-export default LoginForm
+export default RegisterForm
