@@ -1,9 +1,11 @@
 import express from 'express'
 import {register, login} from '../controllers/auth.controller.js'
+import {validate} from '../middlewares/validate.middleware.js'
+import {registerSchema, loginSchema} from '../validations/auth.validation.js'
 
 const router = express.Router()
 
-router.post('/register', register)
-router.post('/login', login)
+router.post('/register', validate(registerSchema), register)
+router.post('/login', validate(loginSchema), login)
 
 export default router
